refactor: src/generated/rest-endpoint-methods-types.ts -> src/generated/types.ts
